requirements.txt
fastapi
uvicorn
opencv-python
pillow
numpy
python-multipart
pytest
pytest-cov
pytest-asyncio
httpx
flake8
pylint